bugfix: FK must have the correct column names and types.
If these are not satisfied, MySQL returns Error 1005 and refers to Error
150 in the error message, which means that a foreign key constraint was
not correctly formed.

+++ sql/pipeline-tables.sql
@@ -270,9 +270,9 @@ CREATE TABLE IF NOT EXISTS panther_annot (
 
 CREATE TABLE IF NOT EXISTS ortholog_goc_metric ( 
   method_link_species_set_id INT NOT NULL,
-  homology_id INT NOT NULL,
-  gene_member_id INT NOT NULL,
-  dnafrag_id INT NOT NULL,
+  homology_id int(10) unsigned NOT NULL,
+  gene_member_id int(10) unsigned NOT NULL,
+  dnafrag_id bigint unsigned NOT NULL,
   goc_score INT NOT NULL, 
   left1 INT,
   left2 INT,
